+++ releases.md
@@ -1,6 +1,6 @@
 # Releases
 
-## Unreleased
+## v2.36.0
 
   - Introduce `Task#wait_all` which recursively waits for all children and self, excepting the current task.
   - Introduce `Task#join` as an alias for `Task#wait` for compatibility with `Thread#join` and similar interfaces.